Solved Blog Posts CRUD - Partially did booking payments

+++ app/Livewire/Admin/EventTypes/Index.php
@@ -2,6 +2,7 @@
 
 namespace App\Livewire\Admin\EventTypes;
 
+use App\Models\Booking;
 use App\Models\EventType;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -17,18 +18,24 @@ class Index extends Component
         'done' => 'render'
     ];
 
-    function delete($id){
-        
+    public function delete($id)
+    {
+
         $event_type = EventType::find($id);
-        $event_type->delete();
+        $event_with_booking = Booking::where('event_type_id', $event_type->id)->exists();
 
-        $this->emit('done', ['success' => "Successfully deleted this event type"]);
+        if ($event_with_booking) {
+            $this->emit('done', ['error' => "Cannot delete " . $event_type->title . " Event Type because it has associated bookings."]);
+        } else {
+            $event_type->delete();
+            $this->emit('done', ['success' => "Successfully deleted " . $event_type->title . " Event Type"]);
+        }
     }
-
     public function render()
     {
-        return view('livewire.admin.event-types.index',[
-            'event_types' => EventType::paginate(5)
+        return view('livewire.admin.event-types.index', [
+            'event_types' => EventType::paginate(10)
         ]);
     }
 }
+
